+++ CherryEngine/imgui.ini
@@ -9,13 +9,13 @@ Size=550,680
 Collapsed=0
 
 [Window][entities]
-Pos=26,13
+Pos=37,16
 Size=323,229
 Collapsed=0
 
 [Window][add CollisionComponent]
-Pos=273,253
-Size=254,117
+Pos=332,261
+Size=136,77
 Collapsed=0
 
 [Window][add DisplayComponent]
