+++ CherryEngine/src/Cherry.h
@@ -1,6 +1,7 @@
 #include "Cherry/CherryI.h"
 #include "Cherry/Logic/SandBoxBase.h"
 #include "Cherry/Logic/EntityComponentSystem.h"
+#include "Cherry/Logic/Systems.h"
 #include "Cherry/Logic/Components.h"
 #include "Cherry/Physics/Shape/Rectangle.h"
 #include "Cherry/Logic/CrEngine.h"
